Peer code review to project 2 from Fan Zhou
Hi Vicki and Eduardo, this is Fan. I am doing code review. Here are some commits and changes I made to your project.
Some praises:
Well-Organized Code Structure: Your repository structure is neat, and it's clear where each component belongs. This makes the project easy to navigate and understand.
Stunning Design Documentation: The design documentation you've provided is not only comprehensive but also visually appealing. The clear, detailed visuals paired with concise explanations make it easy to grasp the intent and flow of the application.

Some feedbacks:
Missing Use of Comments: Your code is not well-commented, which will not be beneficial for future developers (or even yourself) when revisiting the codebase.
Images should be in frontend.
You lost a step to navigate to the directory of your poject.